Bedrock, wool and iron trap door
The bedrock literally only had four colors in it, which is a complete
outlier for this pack, so I added some detail to it.
This wool is a better fit for this more high-contrast direction, I
think. Made some small tweaks to the white and gray ones to make them a
little less rough.